move llama attention to another file
