+++ python/llm/src/bigdl/llm/transformers/model.py
@@ -28,6 +28,14 @@ def save_low_bit(self, *args, **kwargs):
     self.save_pretrained(*args, **kwargs)
 
 
+def convert_forward(m, target_m, new_forward):
+    for _, sub_m in m.named_children():
+        if isinstance(sub_m, target_m):
+            bound_method = new_forward.__get__(sub_m, sub_m.__class__)
+            setattr(sub_m, "forward", bound_method)
+        convert_forward(sub_m, target_m, new_forward)
+
+
 class _BaseAutoModelClass:
 
     HF_MODEL = None
@@ -76,7 +84,7 @@ class _BaseAutoModelClass:
     @classmethod
     def optimize(cls, model):
         from packaging import version
-        from bigdl.llm.transformers.convert import convert_forward, llama_attention_forward_4_31
+        from bigdl.llm.transformers.models.llama import llama_attention_forward_4_31
         trans_version = transformers.__version__
         if version.parse(trans_version) >= version.parse("4.31.0"):
             convert_forward(
